requirements.txt
Flask
Redis
cassandra-driver
Pillow
numpy
tensorflow
logging
